requirements.txt
opencv_python==4.8.1.78
python-dotenv==1.0.0
python-socketio==5.10.0
